requirements.txt
appnope==0.1.0
backcall==0.2.0
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
kiwisolver==1.2.0
numpy==1.19.1
opencv-python==4.4.0.42
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
six==1.15.0
traitlets==5.0.3
wcwidth==0.2.5
